Changed log in Loader when error fetching file by path
I need to think there.
This method uses for fetching most files and it must print errors, bacause it is important to know, when something is broken.
But it is also used to check module existence un source, so if module is located not in first source (for example in pack itself) it will print error as well, but this is normal behaviuor and I think it wont need to be printed...